# Conflicts:
#	.gradle/8.14.3/checksums/checksums.lock
#	.gradle/8.14.3/executionHistory/executionHistory.bin
#	.gradle/8.14.3/executionHistory/executionHistory.lock
#	.gradle/8.14.3/fileHashes/fileHashes.bin
#	.gradle/8.14.3/fileHashes/fileHashes.lock
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.gradle/buildOutputCleanup/outputFiles.bin

+++ FITIDE/src/jvmMain/java/org/ide/FileExplorerController/FileExplorerController.java
@@ -252,9 +252,21 @@ public class FileExplorerController {
 
     private void tryRename(Path pathToFile, String newName) throws UnnableToRenameException {
         File prevFile = new File(pathToFile.toString());
-        File newFile = new File(pathToFile.subpath(0, pathToFile.getNameCount() - 1) + File.separator + newName);
+        File parent = prevFile.getParentFile();
+        if (parent == null) {
+            throw new UnnableToRenameException("No parent directory for: " + pathToFile);
+        }
+
+        File newFile = new File(parent, newName);
+
+        if (newFile.exists()) {
+            throw new UnnableToRenameException("Target already exists: " + newFile.getAbsolutePath());
+        }
+
         if (!prevFile.renameTo(newFile)) {
-            throw new UnnableToRenameException("");
+            throw new UnnableToRenameException(
+                    "Failed to rename " + prevFile.getAbsolutePath() + " to " + newFile.getAbsolutePath()
+            );
         }
     }
